1.0.0
1.添加：发表评论
2.添加：发表评论的评论
3.添加：查询评论列表（支持分页）
4.添加：查询评论列表（具体的评论）
5.添加：查询评论列表（按业务类型）
6.添加：查询评论列表（用户发表的评论）
7.添加：查询评论列表（某一帖子的所有评论）
8.添加：评论点赞
9.添加：评论取消点赞

+++ hy.microservice.post/src/main/java/org/hy/microservice/post/userComment/UserComment.java
@@ -30,14 +30,14 @@ public class UserComment extends BaseViewMode implements TreeObjectNode<UserComm
     /** 帖子编号 */
     private String  postID;     
     
-    /** 帖子内容 */
+    /** 评论内容 */
     private String  comment;  
     
     /** 评分 */
     private Integer score;   
     
     /** 点赞次数 */
-    private Integer goodCount;
+    private Long    goodCount;
     
     /** 评论的回复 */
     private List<UserComment> replys;
@@ -72,7 +72,7 @@ public class UserComment extends BaseViewMode implements TreeObjectNode<UserComm
 
     
     /**
-     * 获取：帖子内容
+     * 获取：评论内容
      */
     public String getComment()
     {
@@ -92,7 +92,7 @@ public class UserComment extends BaseViewMode implements TreeObjectNode<UserComm
     /**
      * 获取：点赞次数
      */
-    public Integer getGoodCount()
+    public Long getGoodCount()
     {
         return goodCount;
     }
@@ -141,7 +141,7 @@ public class UserComment extends BaseViewMode implements TreeObjectNode<UserComm
 
     
     /**
-     * 设置：帖子内容
+     * 设置：评论内容
      * 
      * @param comment 
      */
@@ -167,7 +167,7 @@ public class UserComment extends BaseViewMode implements TreeObjectNode<UserComm
      * 
      * @param goodCount 
      */
-    public void setGoodCount(Integer goodCount)
+    public void setGoodCount(Long goodCount)
     {
         this.goodCount = goodCount;
     }
